+++ beam/beam-core/src/main/java/org/esa/beam/framework/dataio/AbstractProductReader.java
@@ -376,7 +376,7 @@ public abstract class AbstractProductReader implements ProductReader {
     }
 
     /**
-     * Used by the {@link #createTiePointGrid(String, int, int, float, float, float, float, float[]) createTiePointGrid} method in order to determine
+     * Used by the {@link #createTiePointGrid(String, int, int, double, double, double, double, float[]) createTiePointGrid} method in order to determine
      * the discontinuity mode for angle tie-point grids.
      * <p>The default implementation returns {@link TiePointGrid#DISCONT_AT_180} for
      * the names "lon", "long" or "longitude" ignoring letter case,
@@ -410,10 +410,10 @@ public abstract class AbstractProductReader implements ProductReader {
     protected TiePointGrid createTiePointGrid(String gridName,
                                               int gridWidth,
                                               int gridHeight,
-                                              float offsetX,
-                                              float offsetY,
-                                              float subSamplingX,
-                                              float subSamplingY,
+                                              double offsetX,
+                                              double offsetY,
+                                              double subSamplingX,
+                                              double subSamplingY,
                                               float[] tiePoints) {
         final int gridDiscontinutity = getGridDiscontinutity(gridName);
         if (gridDiscontinutity != 0) {
