+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/toolviews/placemark/gcp/GcpTableModel.java
@@ -50,8 +50,8 @@ public class GcpTableModel extends AbstractPlacemarkTableModel {
         Assert.notNull(getProduct());
         final Placemark placemark = getPlacemarkDescriptor().getPlacemarkGroup(getProduct()).get(rowIndex);
 
-        float x = Float.NaN;
-        float y = Float.NaN;
+        double x = Float.NaN;
+        double y = Float.NaN;
 
         final PixelPos pixelPos = placemark.getPixelPos();
         if (pixelPos != null) {
@@ -59,8 +59,8 @@ public class GcpTableModel extends AbstractPlacemarkTableModel {
             y = pixelPos.y;
         }
 
-        float lon = Float.NaN;
-        float lat = Float.NaN;
+        double lon = Float.NaN;
+        double lat = Float.NaN;
 
         final GeoPos geoPos = placemark.getGeoPos();
         if (geoPos != null) {
@@ -68,8 +68,8 @@ public class GcpTableModel extends AbstractPlacemarkTableModel {
             lat = geoPos.lat;
         }
 
-        float dLon = Float.NaN;
-        float dLat = Float.NaN;
+        double dLon = Float.NaN;
+        double dLat = Float.NaN;
 
         final GeoCoding geoCoding = getProduct().getGeoCoding();
